refactor(auth, user): reorganize types and contracts for better structure and clarity
