+++ src/application/user/thunks/createGetCurrentUserThunk.ts
@@ -1,6 +1,7 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
 
-import { GetCurrentUser, User } from "@/domain/user/types";
+import { GetCurrentUser } from "@/domain/user/user.contracts";
+import { User } from "@/domain/user/user.entity";
 import { AppError, AppErrorProps, AppErrorType } from "@/shared/errors/types";
 
 export const createGetCurrentUserThunk = (getCurrentUser: GetCurrentUser) =>
