+++ src/application/user/thunks/createUpdateUserThunk.ts
@@ -1,6 +1,7 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
 
-import { UpdateUser, User } from "@/domain/user/types";
+import { UpdateUser } from "@/domain/user/user.contracts";
+import { User } from "@/domain/user/user.entity";
 import { AppError, AppErrorProps, AppErrorType } from "@/shared/errors/types";
 
 export const createUpdateUserThunk = (updateUser: UpdateUser) =>
