+++ src/domain/auth/auth.payload.ts
@@ -1,0 +1,9 @@
+export interface AuthPayload {
+  email: string;
+  password: string;
+}
+
+export type ChangePasswordPayload = {
+  newPassword: string;
+  oldPassword: string;
+};
