+++ src/domain/user/user.entity.ts
@@ -1,0 +1,6 @@
+export interface User {
+  id: string;
+  username: string;
+  email: string;
+  avatarUrl?: string;
+}
